moved query epoch tracker and module address
