+++ x/stakeibc/client/cli/query_module_address.go
@@ -2,45 +2,6 @@ package cli
 
 import (
 	"strconv"
-
-	"github.com/cosmos/cosmos-sdk/client"
-	"github.com/cosmos/cosmos-sdk/client/flags"
-	"github.com/spf13/cobra"
-
-	"github.com/Stride-Labs/stride/v22/x/stakeibc/types"
 )
 
 var _ = strconv.Itoa(0)
-
-func CmdModuleAddress() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "module-address [name]",
-		Short: "Query module-address",
-		Args:  cobra.ExactArgs(1),
-		RunE: func(cmd *cobra.Command, args []string) (err error) {
-			reqName := args[0]
-
-			clientCtx, err := client.GetClientTxContext(cmd)
-			if err != nil {
-				return err
-			}
-
-			queryClient := types.NewQueryClient(clientCtx)
-
-			params := &types.QueryModuleAddressRequest{
-				Name: reqName,
-			}
-
-			res, err := queryClient.ModuleAddress(cmd.Context(), params)
-			if err != nil {
-				return err
-			}
-
-			return clientCtx.PrintProto(res)
-		},
-	}
-
-	flags.AddQueryFlagsToCmd(cmd)
-
-	return cmd
-}
